+++ acme/Configuration.py
@@ -129,7 +129,6 @@ class Configuration(object):
 				#
 				'cse.security.enableACPChecks'		: config.getboolean('cse.security', 'enableACPChecks', 	fallback=True),
 				'cse.security.adminACPI'			: config.get('cse.security', 'adminACPI', 				fallback='acpAdmin'),
-				'cse.security.defaultACPI'			: config.get('cse.security', 'defaultACPI', 			fallback='acpDefault'),
 				'cse.security.csebaseAccessACPI'	: config.get('cse.security', 'csebaseAccessACPI', 		fallback='acpCSEBaseAccess'),
 				'cse.security.useTLS'				: config.getboolean('cse.security', 'useTLS', 			fallback=False),
 				'cse.security.tlsVersion'			: config.get('cse.security', 'tlsVersion', 				fallback='auto'),